Fixed RPC bugs, settings shows again
