+++ rpc/rpc.go
@@ -42,8 +42,8 @@ func handleRpc(debugPort, serverPort string, plugins *plugins.Plugins, hub *ws.H
 	server.RegisterService(inject.NewInjectService(debugPort, serverPort, plugins, steamPath, authToken, pluginsDir), "InjectService")
 	server.RegisterService(NewPluginsService(plugins), "PluginsService")
 	server.RegisterService(NewIPCService(hub), "IPCService")
-	server.RegisterService(NewAutostartService(dataDir, "crankshaft.service"), "AutostartService")
-	server.RegisterService(NewAutostartService(dataDir, "crankshaft-update.service"), "AutostartService")
+	server.RegisterService(NewSystemdService(dataDir, "crankshaft.service"), "SystemdService")
+	server.RegisterService(NewSystemdService(dataDir, "crankshaft-update.service"), "SystemdService")
 	server.RegisterService(NewExecService(), "ExecService")
 	server.RegisterService(NewStoreService(dataDir), "StoreService")
 	return server
